Move addresspoints model and protocol to shared proejct
