+++ shared/src/main/scala/grasshopper/model/Status.scala
@@ -1,3 +1,3 @@
-package grasshopper.addresspoints.model
+package grasshopper.model
 
 case class Status(status: String, service: String, time: String, host: String)
