+++ shared/src/main/scala/grasshopper/model/addresspoints/AddressInput.scala
@@ -1,3 +1,3 @@
-package grasshopper.addresspoints.model
+package grasshopper.model.addresspoints
 
 case class AddressInput(address: String)
